blocks in chunks are now only stored as ids. In order to reduce RAM usage (~75%)

+++ src/block.c
@@ -1,8 +1,10 @@
 #include <block.h>
 #include <stdlib.h>
 #include <string.h>
+#include <assert.h>
 
-block block_create(int id){
+
+block block_create(uint8_t id){
     block b = {.id = id, .is_solid = true, .is_transparent = false};
     if (id == BLOCK_AIR){
         b.is_solid = false;
